refactor: 헤더 visible 관련 로직을 useHeader 커스텀훅으로 분리
